Switching TC hook to translate git user based on email.
We were using "username", but GH is not supplying that value for Jennifer. No idea why. The only thing consistent is "email", so we're switching to that. No idea which email address gets sent by GH, so at least for George and myself I've got both gmail and AVI-SPL email addresses listed.

This also attempts to find a "pusher" value, which seems to be new in the JSON that GH sends, and represents who pushed. That, to me, seems a better value to use as the user kicking of the build as opposed to the author of the "first" commit we find.