Accept pull request events; change associated user for push events.
Switches "push" handler to use "Sender.Login" info. That info wasn't available in the payload when we first wrote this. It is the most appropriate value to use when determining who to associate the triggered build with.

Updates the lookup used to convert GH login -> TC user to include Mike and Rick, and to no longer be case-sensitive.