Stops triggering builds for Draft PRs.
Hard-coded for now. Could be config-driven, but that would be for
every repo. I would imagine if we ever wanted draft PRs to kick off
builds, we'd do so selectively. So, I didn't bother with a simple
config for this. Will revisit if/when we want to build draft PRs.